+++ CHANGELOG.md
@@ -10,6 +10,11 @@
     [#80](https://github.com/scottlamb/retina/issues/80).
 *   add a new `webcodecs` example that decodes video frames using WebCodecs API.
     This is the absolute lowest-latency way to watch RTSP streams from a browser!
+*   `SetupOptions::strip_inline_parameters`: stripping of inline parameter
+    set NALs (SPS/PPS for H.264, VPS/SPS/PPS for H.265) from `VideoFrame::data`.
+    This is highly recommended for the reasons in the doc comment. It is
+    opt-in for now to avoid surprises on `Cargo.lock` update, but a future
+    major version will strip by default.
 
 ## `v0.4.17` (2026-02-17)